# docs/swagger.yaml
basePath: /
definitions:
...
  todo_internal_models.CreateTaskRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      due_at:
        type: string
      priority:
        enum:
        - 1
        - 2
        - 3
        type: integer
      remind_at:
        type: string
      repeat_end_date:
        type: string
      repeat_interval:
        maximum: 365
        minimum: 1
        type: integer
      repeat_type:
        enum:
        - none
        - daily
        - weekly
        - monthly
        - yearly
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  todo_internal_models.Task:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      due_at:
        type: string
      id:
        type: integer
      priority:
        type: integer
      remind_at:
        type: string
      reminder_sent:
        type: boolean
      reminder_sent_at:
        type: string
      repeat_end_date:
        type: string
      repeat_interval:
        type: integer
      repeat_type:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  todo_internal_models.UpdateTaskRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      due_at:
        type: string
      priority:
        enum:
        - 1
        - 2
        - 3
        type: integer
      remind_at:
        type: string
      repeat_end_date:
        type: string
      repeat_interval:
        maximum: 365
        minimum: 1
        type: integer
      repeat_type:
        enum:
        - none
        - daily
        - weekly
        - monthly
        - yearly
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
  todo_internal_utils.ErrorResponse:
    properties:
      code:
        example: NOT_FOUND
        type: string
      error:
        example: task not found
        type: string
      success:
        example: false
        type: boolean
    type: object
  todo_internal_utils.PageMeta:
    properties:
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  todo_internal_utils.PaginatedResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/todo_internal_utils.PageMeta'
      success:
        example: true
        type: boolean
    type: object
  todo_internal_utils.SuccessResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: 多用户 TODO 任务管理系统，支持用户注册/登录、个人 API Key、多渠道提醒推送
  title: TODO 任务管理系统 API
  version: 2.0.0
paths:
...
  /api/v1/tasks:
    get:
      description: 支持分页查询、按状态/优先级筛选、按关键字搜索、自定义排序
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: created_at
        description: 排序字段
        enum:
        - created_at
        - updated_at
        - due_at
        - priority
        in: query
        name: sort
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 任务状态
        enum:
        - pending
        - completed
        - all
        in: query
        name: status
        type: string
      - description: 优先级筛选
        enum:
        - 1
        - 2
        - 3
        in: query
        name: priority
        type: integer
      - description: 截止时间上限 (2006-01-02 15:04:05)
        in: query
        name: due_before
        type: string
      - description: 截止时间下限 (2006-01-02 15:04:05)
        in: query
        name: due_after
        type: string
      - description: 搜索关键字（标题/描述）
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/todo_internal_models.Task'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取任务列表（分页、筛选、排序）
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: 创建一个新任务，支持设置标题、优先级、截止时间、提醒时间、重复规则等；若设置了提醒时间（remind_at），则必须至少存在一个已启用的通知渠道
      parameters:
      - description: 任务信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建一个新任务
      tags:
      - tasks
  /api/v1/tasks/{id}:
    delete:
      description: 根据任务 ID 永久删除一个任务
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除指定任务
      tags:
      - tasks
    get:
      description: 根据任务 ID 返回任务详情
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 根据 ID 获取任务
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: 部分更新任务字段，只需传入需要修改的字段
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 要更新的字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新任务信息
      tags:
      - tasks
  /api/v1/tasks/{id}/complete:
    patch:
      description: 将任务在已完成和未完成之间切换。如果是重复任务且标记为完成，会自动创建下一次任务
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 切换任务的完成/未完成状态
      tags:
      - tasks
...
securityDefinitions:
  APIKeyAuth:
    in: header
    name: api-key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"